+++ kaioretry/__init__.py
@@ -1,7 +1,8 @@
 """Retry decorator to automatically call a function again on errors"""
 
 import logging
-from typing import Optional, Callable, Awaitable, Union
+from typing import Awaitable
+from collections.abc import Callable
 
 from .types import Exceptions, NonNegative, Number, Jitter, FuncParam, FuncRetVal
 from .context import Context
@@ -16,10 +17,16 @@ RETRY_PARAMS_DOCSTRING = """
     :param exceptions: exceptions classes that will trigger another
         try. Other exceptions raised by the decorated function will
         not trigger a retry. The value of the exceptions parameters
-        can be eiher an :py:class:`Exception` or a :py:class:`tuple`
-        of :py:class:`Exception` (actually, whatever is suitable for
-        an except clause). The default is the :py:class:`Exception`
-        class, which means any error will trigger a new try.
+        can be either an :py:class:`Exception` class or a
+        :py:class:`tuple` of :py:class:`Exception` classes (actually,
+        whatever is suitable for an except clause). The default is the
+        :py:class:`Exception` class, which means any error will
+        trigger a new try.
+
+    :param tries: the maxi number of iterations (a.k.a.: tries,
+        function calls) to perform before exhaustion. A negative value
+        means infinite. 0 is forbidden, since it would mean "don't
+        run" at all.
 
     :param delay: the initial number of seconds to wait between two
         iterations. It must be non-negative. Default is 0.
@@ -31,12 +38,18 @@ RETRY_PARAMS_DOCSTRING = """
     :param jitter: extra seconds added to delay between
         iterations. Default: 0.
 
-    :param jitter: extra seconds added to delay between attempts. default: 0.
-                   fixed if a number, random if a range tuple (min, max)
-
     :param max_delay: the maximum value of delay. default: None (no limit).
 
+    :param min_delay: the minimum value allowed for delay. Cannot be
+        negative. Default is 0.
+
+    :param logger: the :py:class:`logging.Logger` object to which the
+        log messages will be sent to.
+
     :returns: a retry decorator for regular (non-coroutine) functions.
+
+    :raises ValueError: if tries, min_delay or max_delay have incorrect values.
+    :raises TypeError: if jitter is neither a Number or a tuple.
 """